Replace the price allotment Scheme
Earlier, each itemset got a price between 0 and 1 randomly. This means even an itemset with size 1 got between 0 and 1 and an itemset with size 5 got between 0 and 1.
More realistically, the price of itemset should be the sum of itemsets.
Thus, this commit gives random price between 0 and 1 to the items and puts the sum of
itemsets as sum of items it contains.